Assert Tunnelblick is running before sending messages
`tell` starts an application to send it a message. If a user has explicitly quit
Tunnelblick, tunnelblickctl should not start it to check the status, quit it
again, etc.

This ensures the user is in complete control of when Tunnelblick runs.
Specifically, they must either `launch` it or run it as a normal menu
application.